[REQUIRED] properties added to Dash, removed Lash class
