+++ lib/hashie.rb
@@ -6,5 +6,4 @@ module Hashie
   autoload :Mash, 'hashie/mash'
   autoload :Dash, 'hashie/dash'
   autoload :Clash, 'hashie/clash'
-  autoload :Lash, 'hashie/lash'
 end
